--- public/client-index.html
@@ -1,39 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta
-      http-equiv="Content-Security-Policy"
-      content="script-src 'self' 'unsafe-inline';"
-    />
-  </head>
-  <body>
-    <h2>Hello</h2>
-
-    <form id="form">
-      <label>Base Image URL</label>
-      <input type="text" id="base_image_url" name="base_image_url" value="ubuntu:xenial"><br>
-      
-      <label>APT:</label>
-      <textarea id="apt" name="apt">
-wget
-curl
-      </textarea><br>
-      
-      <label>Conda:</label>
-      <textarea id="conda" name="conda"></textarea><br>
-      
-      <label>pip:</label>
-      <textarea id="pip" name="pip"></textarea><br>
-      
-      
-      <div id="output"></div>
-      </form>
-      
-      <button onclick="build_image();" id="build">Build</button>
-
-    <div id="output"></div>
-
-    <script src="client-ipc.js"></script>
-    <script src="client-app.js"></script>
-  </body>
-</html>
