+++ src/BuildImage.jsx
@@ -1,10 +1,21 @@
 import React from 'react';
 import { Layout, Breadcrumb, Form, Input, Button, Row, Col } from 'antd';
+import { send } from './utils/ipcClient';
+
 const { Content } = Layout;
 const { TextArea } = Input;
 
 export default function BuildImage() {
   const placeholder = `one package per line. e.g., \npackage1\npackage2\n`;
+  const [form] = Form.useForm();
+  const onFinish = async (values) => {
+    console.log('form values', values);
+    const result = await send('build-image', values)
+    console.log(result);
+  };
+  const initialValues = {
+    base_image_url: 'ubuntu:xenial',
+  };
   return (
     <Content style={{ margin: '0 16px' }}>
       <Breadcrumb style={{ margin: '16px 0' }}>
@@ -15,8 +26,14 @@ export default function BuildImage() {
         className='site-layout-background'
         style={{ padding: 24, minHeight: 360 }}
       >
-        <Form layout='vertical'>
-          <Form.Item label='Base Image' required>
+        <Form
+          layout='vertical'
+          form={form}
+          name='build_image'
+          initialValues={initialValues}
+          onFinish={onFinish}
+        >
+          <Form.Item label='Base Image' name='base_image_url' required>
             <Input placeholder='e.g., jupyter/base-notebook, jupyter/scipy-notebook' />
           </Form.Item>
           <Form.Item label='Image Name' required>
@@ -24,7 +41,7 @@ export default function BuildImage() {
           </Form.Item>
           <Row gutter={8}>
             <Col span={8}>
-              <Form.Item label='apt'>
+              <Form.Item label='apt' name='apt'>
                 <TextArea
                   allowClear={true}
                   autoSize={{ minRows: 5, maxRows: 5 }}
@@ -33,7 +50,7 @@ export default function BuildImage() {
               </Form.Item>
             </Col>
             <Col span={8}>
-              <Form.Item label='conda'>
+              <Form.Item label='conda' name='conda'>
                 <TextArea
                   allowClear={true}
                   autoSize={{ minRows: 5, maxRows: 5 }}
@@ -42,7 +59,7 @@ export default function BuildImage() {
               </Form.Item>
             </Col>
             <Col span={8}>
-              <Form.Item label='pip'>
+              <Form.Item label='pip' name='pip'>
                 <TextArea
                   allowClear={true}
                   autoSize={{ minRows: 5, maxRows: 5 }}
@@ -52,7 +69,9 @@ export default function BuildImage() {
             </Col>
           </Row>
           <Form.Item style={{ textAlign: 'right' }}>
-            <Button type='primary'>Submit</Button>
+            <Button type='primary' htmlType='submit'>
+              Submit
+            </Button>
             <Button style={{ margin: '0 8px' }}>Reset</Button>
           </Form.Item>
         </Form>
