+++ src/index.jsx
@@ -3,6 +3,9 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
+import * as ipcClient from './utils/ipcClient';
+
+ipcClient.init();
 
 ReactDOM.render(
   <React.StrictMode>
